move: 리소스가 아닌 요청 응답에 따라 dto 를 분리하도록 수정
