+++ src/main/java/roomescape/dto/response/ThemeResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto.theme;
+package roomescape.dto.response;
 
 import roomescape.model.Theme;
