+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,10 +7,10 @@ import roomescape.common.exception.NotFoundException;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.reservation.ReservationRequestDto;
-import roomescape.dto.reservation.ReservationResponseDto;
-import roomescape.dto.reservationtime.ReservationTimeResponseDto;
-import roomescape.dto.theme.ThemeResponseDto;
+import roomescape.dto.request.ReservationRequestDto;
+import roomescape.dto.response.ReservationResponseDto;
+import roomescape.dto.response.ReservationTimeResponseDto;
+import roomescape.dto.response.ThemeResponseDto;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
