+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import roomescape.common.exception.DuplicatedException;
 import roomescape.common.exception.ResourceInUseException;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.dto.reservationtime.AvailableReservationTimeResponseDto;
-import roomescape.dto.reservationtime.ReservationTimeRequestDto;
-import roomescape.dto.reservationtime.ReservationTimeResponseDto;
+import roomescape.dto.response.AvailableReservationTimeResponseDto;
+import roomescape.dto.request.ReservationTimeRequestDto;
+import roomescape.dto.response.ReservationTimeResponseDto;
 import roomescape.model.AvailableReservationTime;
 import roomescape.model.ReservationTime;
