+++ src/test/java/roomescape/dto/ReservationRequestDtoTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import roomescape.dto.reservation.ReservationRequestDto;
+import roomescape.dto.request.ReservationRequestDto;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
