+++ src/test/java/roomescape/dto/ReservationTimeRequestDtoTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import roomescape.dto.reservationtime.ReservationTimeRequestDto;
+import roomescape.dto.request.ReservationTimeRequestDto;
 
 class ReservationTimeRequestDtoTest {
