+++ src/test/java/roomescape/dto/ThemeRequestDtoTest.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dto.theme.ThemeRequestDto;
+import roomescape.dto.request.ThemeRequestDto;
 
 class ThemeRequestDtoTest {
